added user routes and started frontend

+++ backend/controlers/goalControler.js
@@ -1,11 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc  Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
 
   res.status(200).json(goals);
 });
@@ -21,6 +22,7 @@ const setGoals = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -34,6 +36,21 @@ const updateGoals = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+
+  const user = await User.findById(req.user.id);
+
+  // check if user exists
+  if (!user) {
+    res.status(400);
+    throw new Error("User not found");
+  }
+
+  // check if user authorized
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("Not authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -49,6 +66,20 @@ const deleteGoals = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  // check if user exists
+  if (!user) {
+    res.status(400);
+    throw new Error("User not found");
+  }
+
+  // check if user authorized
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("Not authorized");
   }
 
   await Goal.findByIdAndDelete(req.params.id);
